requirements.txt
arrow==1.3.0
attrs==25.3.0
babel==2.17.0
bcrypt==4.2.0
blinker==1.9.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.2
colorzero==2.0
configobj==5.0.9
cryptography==43.0.0
distro==1.9.0
fqdn==1.5.1
gpiod==2.2.0
gpiozero==2.0.1
idna==3.10
isoduration==20.11.0
Jinja2==3.1.6
jsonpatch==1.32
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.12.1
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
packaging==25.0
pycryptodomex==3.20.0
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987==1.3.8
rich==13.9.4
rpds-py==0.21.0
RPi.GPIO==0.7.1a4
smbus2==0.4.3
spidev==3.6
uc-micro-py==1.0.3
uritemplate==4.1.1
urllib3==2.3.0
webcolors==1.13
wheel==0.46.1
flask
smbus2
RPi.bme280
adafruit-blinka
adafruit-circuitpython-ads1x15
RPLCD
